% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaart.R
\name{kaart}
\alias{kaart}
\title{Maakondade kaart}
\usage{
kaart(
  df,
  fill,
  labels,
  title = "",
  legend.title = "",
  textsize = 7,
  annotation = "kaart: Maa-amet, 2021"
)
}
\arguments{
\item{df}{andmetabeli nimi}

\item{fill}{tunnuse nimi andmetabelist (see, mis l채heb v채rvidena kaardile)}

\item{labels}{tunnuse nimi andmetabelist (tekst, mis l채heb kaardile)}

\item{title}{joonise nimi}

\item{legend.title}{legendi nimi}

\item{textsize}{teksti suurus kaardil (nb! 체hikute kohta vt https://stackoverflow.com/questions/25061822/ggplot-geom-text-font-size-control)}

\item{annotation}{annotatsioon all paremal servas}
}
\value{
a ggplot
}
\description{
Maakondade kaart (siia pikem kirjeldus!)
}
\details{
tba
}
