% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaart.R
\name{kaart2}
\alias{kaart2}
\title{Maakondade kaart Tallinna ja Tartuga}
\usage{
kaart2(
  df,
  fill,
  labels,
  title = "",
  legend.title = "",
  textsize = 7,
  annotation = "kaart: Maa-amet, 2021"
)
}
\arguments{
\item{df}{andmetabeli nimi}

\item{fill}{tunnuse nimi andmetabelist (see, mis läheb värvidena kaardile)}

\item{labels}{praegu kasutamata}

\item{title}{joonise nimi}

\item{legend.title}{legendi nimi}

\item{textsize}{teksti suurus kaardil (nb! ühikute kohta vt https://stackoverflow.com/questions/25061822/ggplot-geom-text-font-size-control)}

\item{annotation}{annotatsioon all paremal servas}
}
\value{
a ggplot
}
\description{
Maakondade kaart Tallinna ja Tartuga (siia pikem kirjeldus?!)
}
\details{
tba
}
