% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaart.R
\name{kaart}
\alias{kaart}
\title{Maakondade kaart}
\usage{
kaart(
  df,
  fill,
  labels,
  title = "",
  legend.title = "",
  annotation = "kaart: Maa-amet, 2021"
)
}
\arguments{
\item{df}{andmetabeli nimi}

\item{fill}{tunnuse nimi andmetabelist (see, mis läheb värvidena kaardile)}

\item{labels}{tunnuse nimi andmetabelist (tekst, mis läheb kaardile)}

\item{title}{joonise nimi}

\item{legend.title}{legendi nimi}

\item{annotation}{annotatsioon all paremal servas}
}
\value{
a ggplot
}
\description{
Maakondade kaart (siia pikem kirjeldus!)
}
\details{
tba
}
